feat: extract and use suspense to reduce waterfall
