+++ src/app/search/utils/ResultItemsSection.tsx
@@ -3,13 +3,12 @@
 import { Activity } from '@action/activity';
 import { SkeletonEventCard } from '@components/EventCard';
 import { SkelotonSearchResultEventCard } from '@components/EventCard/SearchResultEventCard';
-import WithErrorBoundaryAndSuspense from '@components/hoc/WithErrorBoundaryAndSuspense';
 import useMediaQuery from '@hooks/use-media-query';
 import { useSearchParams } from 'next/navigation';
 import { H4 } from '../../../components/ui/typography';
 import ResultItems from './ResultItems';
 
-const SkelotonItems = ({ isMobile }: { isMobile: boolean }) => {
+export const SkelotonItems = ({ isMobile }: { isMobile: boolean }) => {
   if (isMobile) {
     return (
       <>
@@ -51,16 +50,12 @@ const ResultItemsSection = ({
           「{keyword}」找到{total}個活動
         </H4>
       </div>
-      <WithErrorBoundaryAndSuspense
-        loadingFallback={<SkelotonItems isMobile={isMobile} />}
-      >
-        <ResultItems
-          results={results}
-          setCenterId={setCenterId}
-          isMobile={isMobile}
-          className="lg:max-w-[53.5rem] flex w-full flex-col gap-y-12"
-        />
-      </WithErrorBoundaryAndSuspense>
+      <ResultItems
+        results={results}
+        setCenterId={setCenterId}
+        isMobile={isMobile}
+        className="lg:max-w-[53.5rem] flex w-full flex-col gap-y-12"
+      />
     </>
   );
 };
