+++ src/app/search/utils/ResultMapSection.tsx
@@ -1,11 +1,9 @@
 'use client';
 
 import { Activity } from '@action/activity';
-import WithErrorBoundaryAndSuspense from '@components/hoc/WithErrorBoundaryAndSuspense';
 import { useSearchProvider } from '@components/search/SearchBar/SearchProvider';
 import { useSearchView } from '@components/search/SearchBar/SearchViewProvider';
 import SearchMapSection from '@components/search/SearchMapSection';
-import { Skeleton } from '@components/ui/skeleton';
 import { useMemo } from 'react';
 
 type ResultMapSectionProps = {
@@ -54,29 +52,18 @@ const ResultMapSection = ({ activities = [] }: ResultMapSectionProps) => {
   }, [activities]);
 
   return (
-    <WithErrorBoundaryAndSuspense
-      loadingFallback={
-        <div className="h-[400px] w-full space-y-2">
-          <Skeleton className="h-full w-full rounded-lg" />
-          <div className="flex justify-between">
-            <Skeleton className="h-4 w-[100px]" />
-            <Skeleton className="h-4 w-[60px]" />
-          </div>
-          <Skeleton className="h-4 w-[80%]" />
-        </div>
-      }
-    >
+    <>
       {isMobile && currentShow === 'map' && (
         <div className="h-fit w-full">
           <SearchMapSection centerId={centerId} markers={mapMarkers} />
         </div>
       )}
       {!isMobile && (
-        <div className="sticky top-0 h-fit w-full max-w-[26rem]">
+        <div className="sticky top-0 h-fit w-full xl:max-w-[26rem]">
           <SearchMapSection centerId={centerId} markers={mapMarkers} />
         </div>
       )}
-    </WithErrorBoundaryAndSuspense>
+    </>
   );
 };
