+++ src/app/search/utils/ResultsPagination.tsx
@@ -6,15 +6,14 @@ import Pagination, {
   PaginationPrev,
   generatePaginationItems,
 } from '@components/Pagination';
+import { updateQueryString } from '@components/search/SearchBar';
+import { useSearchProvider } from '@components/search/SearchBar/SearchProvider';
 import { cva } from 'class-variance-authority';
+import Link from 'next/link';
 
 export type ResultsPaginationProps = {
   totalPage: number;
   currentPage: number;
-  isMobile?: boolean;
-  onClickPrev?: (currentPage: number) => void;
-  onClickNext?: (currentPage: number) => void;
-  onPageChange?: (page: number) => void;
 };
 
 const paginationStepperStyles = cva('flex gap-4 py-12', {
@@ -29,29 +28,29 @@ const paginationStepperStyles = cva('flex gap-4 py-12', {
 const ResultsPagination = ({
   totalPage,
   currentPage,
-  isMobile = false,
-  onClickPrev,
-  onClickNext,
-  onPageChange,
 }: ResultsPaginationProps) => {
-  const handlePrevClick = () => {
-    onClickPrev?.(currentPage);
-  };
-
-  const handleNextClick = () => {
-    onClickNext?.(currentPage);
+  const { isMobile } = useSearchProvider();
+  const handlePageChange = (newPage: number) => {
+    const updatedQuery = updateQueryString({
+      page: newPage.toString(),
+    });
+    return `/search?${updatedQuery}`;
   };
 
   const renderPaginationItems = () => {
     return generatePaginationItems(currentPage, totalPage).map((item) => {
       if (item.type === PaginationItem) {
         return (
-          <PaginationItem
+          <Link
             key={item.key}
-            page={item.props.page}
-            isCurrent={item.props.isCurrent}
-            onClick={() => onPageChange?.(item.props.page)}
-          />
+            href={handlePageChange(item.props.page)}
+            passHref
+          >
+            <PaginationItem
+              page={item.props.page}
+              isCurrent={item.props.isCurrent}
+            />
+          </Link>
         );
       }
       return item;
@@ -59,26 +58,27 @@ const ResultsPagination = ({
   };
 
   return (
-    <Pagination
-      currentPage={currentPage}
-      totalPage={totalPage}
-      onClickPrev={handlePrevClick}
-      onClickNext={handleNextClick}
-      asChild
-    >
+    <Pagination currentPage={currentPage} totalPage={totalPage} asChild>
       <div
         id="pagination-stepper"
         className={paginationStepperStyles({ isMobile })}
       >
-        <PaginationPrev
-          className={isMobile ? 'border-[1px]' : ''}
-          iconClassName={isMobile ? 'size-12' : ''}
-        />
+        <Link href={handlePageChange(Math.max(currentPage - 1, 1))} passHref>
+          <PaginationPrev
+            className={isMobile ? 'border-[1px]' : ''}
+            iconClassName={isMobile ? 'size-12' : ''}
+          />
+        </Link>
         {isMobile ? null : renderPaginationItems()}
-        <PaginationNext
-          className={isMobile ? 'border-[1px]' : ''}
-          iconClassName={isMobile ? 'size-12' : ''}
-        />
+        <Link
+          href={handlePageChange(Math.min(currentPage + 1, totalPage))}
+          passHref
+        >
+          <PaginationNext
+            className={isMobile ? 'border-[1px]' : ''}
+            iconClassName={isMobile ? 'size-12' : ''}
+          />
+        </Link>
       </div>
     </Pagination>
   );
